[ui-char-groups] move resolveChars from videoproof-contextual-template.mjs
