+++ lib/js/components/actors/videoproof-contextual-template.mjs
@@ -19,10 +19,6 @@
 
 import {
     getCharsFromCharGroups
-  , getExtendedChars
-  , getCharsForSelectUI
-  , getCharsForKey
-  , getSelectedChars
 } from '../ui-char-groups.mjs';
 
 import {
@@ -341,34 +337,6 @@ function* pairProductGen(outer, inner) {
     for(const o of outer)
         for(const i of inner)
             yield [o, i];
-}
-
-/**
- * Resolve chars from charGroupsData + fonts for a given charGroup config.
- *
- * @param {Object} charGroupsData
- * @param {Array} fonts
- * @param {string} charGroupOption — e.g. 'Latin.Lowercase', 'custom', 'all-gid', etc.
- * @param {boolean} showExtended
- * @param {string} [customText='']
- * @param {string} [customSeparator='']
- * @returns {Array<string>}
- */
-export function resolveChars(charGroupsData, fonts, charGroupOption, showExtended
-                            , customText='', customSeparator='') {
-    let charsData;
-    if(charGroupOption === 'custom') {
-        const chars = customSeparator !== ''
-                    ? customText.split(customSeparator)
-                    : [...customText]
-          , extendedChars = getExtendedChars(charGroupsData, chars)
-          ;
-        charsData = { chars, extendedChars, hasExtended: extendedChars.size > 0 };
-    }
-    else {
-        charsData = getCharsForSelectUI(charGroupsData, fonts, charGroupOption);
-    }
-    return getSelectedChars(charsData, showExtended);
 }
 
 /**
