+++ lib/js/components/actors/videoproof-contextual.mjs
@@ -35,8 +35,11 @@ export {
 } from './videoproof-contextual-models.mjs';
 
 import {
+    resolveChars
+} from '../ui-char-groups.mjs';
+
+import {
     compileTemplateFromPropertyValuesMap
-  , resolveChars
   , generateWords
 } from './videoproof-contextual-template.mjs';
