+++ lib/js/components/ui-char-groups.mjs
@@ -102,7 +102,7 @@ export function getCharsFromCharGroups(charGroups, keyPath) {
         // doesn't feel like the correct thing to do.
         extendedChars = getExtendedChars(charGroups, chars);
         if(target._extended.length)
-            // getExtendedChars returns a nre Map, so that we can do this
+            // getExtendedChars returns a new Map, so that we can do this
             // however, I wonfer if we should cache this as well?
             extendedChars.set(null, [...target._extended]);
     }
@@ -382,6 +382,35 @@ export function getSelectedChars(charsData, showExtended) {
         return selectedChars;
     }
     return Array.from(charsData.chars);
+}
+
+/**
+ * Resolve chars from charGroupsData + fonts for a given charGroup config.
+ *
+ * @param {Object} charGroupsData
+ * @param {Array} fonts
+ * @param {string} charGroupOption — e.g. 'Latin.Lowercase', 'custom', 'all-gid', etc.
+ * @param {boolean} showExtended
+ * @param {string} [customText='']
+ * @param {string} [customSeparator='']
+ * @returns {Array<string>}
+ */
+// TODO: should be used in videoproof-array as well, as it has originated there
+export function resolveChars(charGroupsData, fonts, charGroupOption, showExtended
+                            , customText='', customSeparator='') {
+    let charsData;
+    if(charGroupOption === 'custom') {
+        const chars = customSeparator !== ''
+                    ? customText.split(customSeparator)
+                    : [...customText]
+          , extendedChars = getExtendedChars(charGroupsData, chars)
+          ;
+        charsData = { chars, extendedChars, hasExtended: extendedChars.size > 0 };
+    }
+    else {
+        charsData = getCharsForSelectUI(charGroupsData, fonts, charGroupOption);
+    }
+    return getSelectedChars(charsData, showExtended);
 }
 
 /**
